CLN: refactored the segmentation routines
