+++ pyfibre/model/tools/segmentation.py
@@ -87,13 +87,11 @@ def rgb_segmentation(
     return sorted_cells, sorted_fibres
 
 
-def cell_segmentation(
-        multi_image, fibre_networks, scale=1.0, pl_analysis=False):
+def cell_segmentation(multi_image, fibre_networks,
+                      scale=1.0, pl_analysis=False):
 
-    fibre_segments = [
-        fibre_network.segment
-        for fibre_network in fibre_networks
-    ]
+    fibre_segments = [fibre_network.segment
+                      for fibre_network in fibre_networks]
 
     if pl_analysis:
 
@@ -102,6 +100,7 @@ def cell_segmentation(
         fibre_net_binary = segments_to_binary(
             fibre_segments, multi_image.shape
         )
+
         fibre_filter = np.where(fibre_net_binary, 2, 0.25)
         fibre_filter = gaussian_filter(fibre_filter, 0.5)
 
@@ -117,7 +116,8 @@ def cell_segmentation(
         fibre_col_binary = segments_to_binary(
             fibre_col_seg,  multi_image.shape
         )
-        fibre_col_binary = binary_dilation(fibre_col_binary, iterations=2)
+        fibre_col_binary = binary_dilation(
+            fibre_col_binary, iterations=2)
         fibre_col_binary = binary_closing(fibre_col_binary)
 
         # Generate a filter for the SHG image that combines information
@@ -125,15 +125,11 @@ def cell_segmentation(
         fibre_binary = mean_binary(
             np.array([fibre_net_binary, fibre_col_binary]),
             multi_image.shg_image,
-            min_intensity=0.13)
+            min_intensity=0.10)
 
         # Create a new set of segments for each cell region
-        cell_segments = binary_to_segments(
-            ~fibre_binary, multi_image.pl_image, 250, 0.01)
-
-        cells = [Cell(segment=cell_segment,
-                      image=multi_image.pl_image)
-                 for cell_segment in cell_segments]
+        image = multi_image.pl_image
+        cell_binary = np.where(fibre_binary, 0, 1)
 
     else:
         # Create a filter for the PL image that corresponds
@@ -142,15 +138,16 @@ def cell_segmentation(
         fibre_binary = segments_to_binary(
             fibre_segments, multi_image.shape
         )
-        cell_binary = ~fibre_binary
 
-        # Segment the PL image using this filter
-        cell_segments = binary_to_segments(
-            cell_binary, multi_image.shg_image, 250, 0.01
-        )
+        image = multi_image.shg_image
+        cell_binary = np.where(fibre_binary, 0, 1)
 
-        cells = [Cell(segment=cell_segment,
-                      image=multi_image.shg_image)
-                 for cell_segment in cell_segments]
+    # Segment the image using this filter
+    cell_segments = binary_to_segments(
+        cell_binary, image, 200, 0.001
+    )
+
+    cells = [Cell(segment=cell_segment, image=image)
+             for cell_segment in cell_segments]
 
     return cells
